chore(cosmos): refactor,  move base client, move submodule out of chain in configs
