+++ modules/cosmos/src/config/module.ts
@@ -2,6 +2,7 @@ import { Account } from "@firewatch/core/account";
 import { ModuleConfig } from "@firewatch/core/module";
 import { Chain } from "@firewatch/core/chain";
 import { BankModule } from "../modules/bank/module";
+import { SlashingModule } from "../modules/slashing/module";
 
 export interface CosmosModuleConfig extends Omit<ModuleConfig<Chain, Account>, "network" | "accounts"> {
     chain: Chain;
@@ -10,4 +11,5 @@ export interface CosmosModuleConfig extends Omit<ModuleConfig<Chain, Account>, "
 
 export type SubModulesObject = {
     bank: BankModule;
+    slashing: SlashingModule;
 };
