+++ modules/cosmos/src/modules/slashing/module.ts
@@ -1,6 +1,4 @@
-import { CosmosModule } from "../../models/module";
-
-export interface SlashingModule extends CosmosModule {
-    downtimeFraction: string;
-    doubleSignerFraction: string;
+export interface SlashingModule {
+    slashDowntimeFraction: string;
+    slashDoubleSignFraction: string;
 }
